requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
quart>=0.19.4
